wire bus for console purpose
